Rename: `application`, `domain`, `infrastructure`, `presentation` 구조 적용
