+++ src/main/java/subway/Line/application/LineService.java
@@ -1,4 +1,4 @@
-package subway;
+package subway.Line.application;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -7,6 +7,14 @@ import java.util.Objects;
 import java.util.stream.Collectors;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import subway.Line.domain.Line;
+import subway.Line.presentation.dto.LineRequest;
+import subway.Line.presentation.dto.LineResponse;
+import subway.Line.infrastructure.LineRepository;
+import subway.Station.domain.Station;
+import subway.Station.infrastructure.StationRepository;
+import subway.Station.presentation.dto.StationResponse;
+import subway.Station.domain.Stations;
 
 @Service
 @Transactional(readOnly = true)
