+++ src/main/java/subway/Line/domain/Line.java
@@ -1,7 +1,6 @@
-package subway;
+package subway.Line.domain;
 
 import java.util.List;
-import java.util.Objects;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
