+++ src/main/java/subway/Line/infrastructure/LineRepository.java
@@ -1,6 +1,7 @@
-package subway;
+package subway.Line.infrastructure;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import subway.Line.domain.Line;
 
 public interface LineRepository extends JpaRepository<Line, Long> {
