+++ src/main/java/subway/Line/presentation/LineController.java
@@ -1,4 +1,4 @@
-package subway;
+package subway.Line.presentation;
 
 import java.net.URI;
 import java.util.List;
@@ -9,6 +9,9 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
+import subway.Line.presentation.dto.LineRequest;
+import subway.Line.presentation.dto.LineResponse;
+import subway.Line.application.LineService;
 
 @RestController
 public class LineController {
