+++ src/main/java/subway/Line/presentation/dto/LineRequest.java
@@ -1,4 +1,4 @@
-package subway;
+package subway.Line.presentation.dto;
 
 public class LineRequest {
