+++ src/main/java/subway/Station/application/StationService.java
@@ -1,10 +1,14 @@
-package subway;
+package subway.Station.application;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.stream.Collectors;
+import subway.Station.domain.Station;
+import subway.Station.infrastructure.StationRepository;
+import subway.Station.presentation.dto.StationRequest;
+import subway.Station.presentation.dto.StationResponse;
 
 @Service
 @Transactional(readOnly = true)
