+++ src/main/java/subway/Station/infrastructure/StationRepository.java
@@ -1,7 +1,8 @@
-package subway;
+package subway.Station.infrastructure;
 
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
+import subway.Station.domain.Station;
 
 public interface StationRepository extends JpaRepository<Station, Long> {
